@@ -32,17 +32,17 @@
   "dependencies": {
     "@amcharts/amcharts4": "^4.10.22",
     "@amcharts/amcharts4-geodata": "^4.1.21",
-    "@angular/animations": "~13.3.11",
+    "@angular/animations": "^14.2.0",
     "@angular/cdk": "^12.2.13",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "^13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/localize": "^13.3.11",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/localize": "^14.2.0",
     "@angular/material": "^12.2.13",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
     "html2canvas": "^1.3.2",
     "pdfmake": "^0.1.72",
     "rxjs": "^6.6.7",
@@ -50,9 +50,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular/cli": "^13.3.9",
-    "@angular/compiler-cli": "^13.3.11",
+    "@angular-devkit/build-angular": "^14.2.1",
+    "@angular/cli": "^14.2.1",
+    "@angular/compiler-cli": "^14.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.20.26",
